requirements.txt
scipy
einops
pyyaml
huggingface_hub
encodec
matplotlib
transformers>=4.46.1
pytorch-lightning
tensorboardX
soundfile
numpy
jsonargparse
torchcrepe
librosa
pesq
inflect
loguru
polars
natsort
tqdm
requests
sounddevice
mecab-python3
unidic-lite
uroman

llama-cpp-python
outetts-0.2.0-py3-none-any.whl
